test(weather): Move ViewModel tests to feature-specific directory
- Relocated ViewModel-related test files to test/features/weather/view_model/ for modular structure.
- Updated relative imports to match new mock locations.
