+++ test/features/weather/view_model/error_view_model_test.dart
