+++ test/features/weather/view_model/providers/custom_mock_city_search_view_model_provider.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:weather_app/features/weather/view_model/city_search_state.dart';
 
-import '../../features/weather/mocks/custom_mock_city_search_view_model.dart';
+import '../../mocks/custom_mock_city_search_view_model.dart';
 
 /// [CustomMockCitySearchViewModel]を提供する[NotifierProvider]
 final customMockCitySearchViewModelProvider =
